+++ castle/cms/tiles/existing.py
@@ -1,14 +1,43 @@
+from castle.cms.tiles.base import DisplayTypeTileMixin
 from castle.cms.tiles.content import ContentTile
 from castle.cms.tiles.content import IContentTileSchema
+from castle.cms.tiles.views import BaseTileView
+from castle.cms.tiles.views import TileViewsSource
 from castle.cms.widgets import FocalPointSelectFieldWidget
+from castle.cms.widgets import PreviewSelectFieldWidget
 from plone.autoform import directives as form
+from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from z3c.form.browser.checkbox import CheckBoxFieldWidget
 from zope import schema
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
 
 
-class ExistingTile(ContentTile):
+DISPLAY_TYPE_KEY = 'existing'
+
+
+class BasicView(BaseTileView):
+    name = 'basic'
+    preview = ''
+    order = 0
+    index = ViewPageTemplateFile('templates/existing/basic.pt')
+    tile_name = DISPLAY_TYPE_KEY
+
+
+class BackgroundImageView(BaseTileView):
+    name = 'backgroundimage'
+    preview = ''
+    order = 1
+    index = ViewPageTemplateFile('templates/existing/backgroundimage.pt')
+    tile_name = DISPLAY_TYPE_KEY
+
+
+class ExistingTile(ContentTile, DisplayTypeTileMixin):
+
+    display_type_name = DISPLAY_TYPE_KEY
+    display_type_default = 'basic'
+    display_type_fallback_view = BasicView
+
     def joiner(self, val, joiner=', '):
         if not val:
             return ''
@@ -73,6 +102,14 @@ class IExistingTileSchema(IContentTileSchema):
             SimpleTerm('backgroundimage', 'backgroundimage', u'Background Image')
         ]),
         default='basic'
+    )
+
+    form.widget('display_type', PreviewSelectFieldWidget,
+                tile_name=DISPLAY_TYPE_KEY)
+    display_type = schema.Choice(
+        title=u"Display Type",
+        source=TileViewsSource(DISPLAY_TYPE_KEY),
+        default='default'
     )
 
     form.widget(override_focal_point=FocalPointSelectFieldWidget)