+++ castle/cms/upgrades/__init__.py
@@ -1,0 +1,1 @@
+PROFILE_ID = 'profile-castle.cms:default'